CC/fastapi/requirements.txt
fastapi
uvicorn
tensorflow
pillow
bcrypt
pydantic
numpy
firebase-admin